Initally added a z-index to the css for the searchbar, but it took out the background, so it's gone.
Added the backbone of the expanded view into the JS file. [Relevant] data from searches will be stored in an array, and all geospots will have an id related to the index in the array. Each search will have an expand function that will display the data properly. Any better ideas for how to store data are welcome. I didn't find much when I searched beyond 'jQuery to the rescue'.